refactor: rename discovery service modules and update imports
